fix: LEAP-443: Select only one annotation on init (#1640)
### Main change

**Only one annotation is selected during task opening.**

We have the concept of selecting annotation, when we not just assign it as current one, but also update data in tags, trigger external event, load annotation history, setup hotkeys, set initial values.

Previosly it was required to select every annotation to do some extra work in regions/tags inside it because of some quirky legacy code. I fixed what I found so far and ran all possible tests — they all are green, so consider this change mostly safe.

Benefits:
- improved performance as only one annotation is selected, reducing number of calculations and renders
- no excess network requests to retrieve annotation history for every of these annotations
- more clean code
- one more step to remove one of the oldest legacy concept of `states`: it's not used in init anymore

Things changed:
- `STORE_INIT_OK` global var added to catch tags that can't properly work with new system (edge cases we missed)
- apparently we have only one unit test for tags — Ranker; it's securely mocked

Concerns:
- `states` in Video tag select only *Labels controls, but changed code is in `fixBrokenAnnotation` which will do nothing to Video anyway
- in potential missed edge case app will just crash
- we could potentially check that we are inside init process by some flag, right?

### Other changes
- removed Predictions panel and related code like showAllPredictions
- removed panels from App and down the stream — we don't use them anywhere (3rd party LSF can be affected, but not sure anyone using panels there)
- converted `LabelStudio.js` to TypeScript
- fixed bunch of annoying react warnings (non-html props, wrong values, missed refs)
- started to reorganize our "envs" to get rid of them at some point; for now it's `getRoot()` removed from there
- removed `SidebarPage` as it was useless, plus removed multi-tabs support from `SidebarTabs` as `panels` were not used; plus that's all only for the old interface which is about to be removed
- removed `hydrated` as it's not used anymore after https://github.com/HumanSignal/label-studio-frontend/pull/1166
- changed order of annotations to original one without reversing; thus order will always be the same, fixing the issue when list was reversed again after task reload; visual order is defined in `AnnotationsCarousel` anyway, so no changes in UX
- fixed annotations order in View All to be the same as in annotation tabs



* Remove View All Predictions and anything related

It's not used anymore for a while

* Move LabelStudio to TS, improve a little

- Type drafts for main LSF class
- getRootElement moved from env
- Fixed methods signatures
- Fixed typo in method name

Most of the imports have TS errors because they are not typed;
let it be like this for now.

* Simplify SidebarTabs and remove unused panels param

`panels` were not used anywhere, and this is part of old interface,
that will be removed soon.
+ fix global Htx typing
+ fix conditions in App

* Remove `hydrated`, we don't need it after #1166

* Fix some annoying react warnings

* Don't select every annotation on init

Also don't call any unneccessary methods for render.
Just init regions.

* Add flags to test correct behavior of init

* Fix Relation init

* fix our fixBrokenAnnotation() method

Don't use states(), we already have simple toNames

* Some other small fixes

React warnings, simplier code

* Fix merged results - they can also use toNames

This is needed only for Video rectangles

* Temporary fix for unit tests

We don't have a real init, so overpass it

* Fix history reinit for some cases

Should fix failed e2e with TimeSeries, Audio and Video

* Fix annotation selected by default: should be 1st one

Annotations added to the start of the list, so the last one added
is the first one in the list and is selected by default

* Remove unused Annotation#update model prop

* Add feature flag to cover annotation select

* Fix FF name

* Add a description for this FF

* Add doc with lot of investigations

* Tiny fix for panel classnames

no extra whitespaces

* Fix from #1665

* Fix typos in LSF.init.md

* Trigger LS build

* Fix typo to trigger LS build again

* Move simpleInit to volatile

This allows to check FF indirectly in other apps

* Add more info about TopBar

* Add couple of deprecation messages

* Stop reversing annotations list

Reversing the list caused problems when task is reloaded and list is reversed again.
The visual order is defined in AnnotationsCarousel, so technical order is not that important.
View All used technical order which can cause discomfort when you see different order
in different places. So that was also fixed.
Everything is behind the same FF, without it the order is an old one, reversed.

* Move annotations ordering to utilities; sort imports

* Fix annotation selected after View All

---------

Co-authored-by: hlomzik <hlomzik@users.noreply.github.com>